feat(plugin): ✨ Refactor blog plugin structure and enhance file handling
