+++ apps/web/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "init": "vitnode init",
-    "test": "vitnode test",
+    "prepare": "vitnode prepare",
     "drizzle-kit": "drizzle-kit",
     "db:push": "drizzle-kit push",
     "db:migrate": "drizzle-kit up && drizzle-kit generate && drizzle-kit migrate",
