--- apps/web/src/app/[locale]/(main)/(plugins)/(blog)/blog/layout.tsx
@@ -1,5 +1,0 @@
-import { TestLayout } from 'vitnode-blog/views/test/layout';
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <TestLayout>{children}</TestLayout>;
-}
