--- apps/web/src/app/[locale]/(main)/(plugins)/(blog)/blog/test/page.tsx
@@ -1,9 +1,0 @@
-import { Link } from 'vitnode/lib/navigation';
-
-export default function Page() {
-  return (
-    <>
-      <Link href="/blog">Go to blog</Link>
-    </>
-  );
-}
