--- apps/web/src/app/[locale]/(main)/(vitnode)/login/sso/[providerId]/page.tsx
@@ -1,15 +1,0 @@
-import { CallbackSSOView } from 'vitnode/views/auth/sso/callback/callback-sso-view';
-
-export default async function Page({
-  params,
-  searchParams,
-}: {
-  params: Promise<{ providerId: string }>;
-  searchParams: Promise<Record<string, string>>;
-}) {
-  const { providerId } = await params;
-
-  return (
-    <CallbackSSOView providerId={providerId} searchParams={searchParams} />
-  );
-}
