--- apps/web/src/app/[locale]/admin/(auth)/(plugins)/blog/categories/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div>Categories from admin - blog plugin</div>;
-}
