--- apps/web/src/app/[locale]/admin/(auth)/(vitnode)/core/page.tsx
@@ -1,5 +1,0 @@
-import { DashboardAdminView } from 'vitnode/views/admin/views/core/dashboard/dashboard-admin-view';
-
-export default function Page() {
-  return <DashboardAdminView />;
-}
