--- apps/web/src/app/[locale]/admin/(auth)/(vitnode)/core/test/page.tsx
@@ -1,5 +1,0 @@
-import { TestView } from 'vitnode/views/admin/views/core/test';
-
-export default function Page() {
-  return <TestView />;
-}
