--- apps/web/src/app/[locale]/admin/(auth)/(vitnode)/core/users/page.tsx
@@ -1,7 +1,0 @@
-import { UsersAdminView } from 'vitnode/views/admin/views/core/users/users-admin-view';
-
-export default function Page(
-  props: React.ComponentProps<typeof UsersAdminView>,
-) {
-  return <UsersAdminView {...props} />;
-}
