+++ packages/vitnode/scripts/get-config.ts
@@ -2,14 +2,16 @@
 import { join } from 'path';
 import { pathToFileURL } from 'url';
 
-export const getConfig = async () => {
+import type { VitNodeConfig } from '../src/vitnode.config';
+
+export const getConfig = async (): Promise<VitNodeConfig> => {
   const configPath = join(process.cwd(), 'src', 'vitnode.config.ts');
   try {
     const configUrl = pathToFileURL(configPath).href;
     const loaded = await import(configUrl);
+    const config = loaded.vitNodeConfig;
 
-    const config = loaded.vitNodeConfig ?? loaded.default;
-    console.log('Config metadata title:', config.metadata.title);
+    return config;
   } catch (error) {
     console.error('Failed to load config:', error);
     process.exit(1);
