+++ packages/vitnode/scripts/scripts.ts
@@ -3,10 +3,9 @@
 import { existsSync } from 'fs';
 import { join } from 'path';
 
-import { getConfig } from './get-config.js';
 import { processPlugin } from './plugin.js';
 import { prepareDatabase } from './prepare-database.js';
-import { prepareFiles } from './prepare-files.js';
+import { prepareFiles } from './prepare/prepare-files.js';
 
 const initMessage = '\x1b[34m[VitNode]\x1b[0m';
 const getPluginsPath = () => {
@@ -21,20 +20,27 @@ const getPluginsPath = () => {
   return pluginsPath;
 };
 
-if (process.argv[2] === 'prepare') {
-  void prepareFiles({ pluginsPath: getPluginsPath(), initMessage });
-} else if (process.argv[2] === 'init') {
-  void prepareDatabase({ initMessage });
-} else if (
-  process.argv[2] === 'plugin' &&
-  (process.argv[3] === '--w' || process.argv[3] === '--watch')
-) {
-  processPlugin({ initMessage });
-} else if (process.argv[2] === 'test') {
-  void getConfig();
-} else {
-  console.log(
-    `${initMessage} \x1b[31mCommand not found: "${process.argv[2] ?? ''}"\x1b[0m`,
-  );
-  process.exit(1);
+const command = process.argv[2];
+const flag = process.argv[3];
+
+switch (command) {
+  case 'init':
+    void prepareDatabase({ initMessage });
+    break;
+
+  case 'plugin':
+    if (flag === '--w' || flag === '--watch') {
+      processPlugin({ initMessage });
+    }
+    break;
+
+  case 'prepare':
+    void prepareFiles({ pluginsPath: getPluginsPath(), initMessage });
+    break;
+
+  default:
+    console.log(
+      `${initMessage} \x1b[31mCommand not found: "${command ?? ''}"\x1b[0m`,
+    );
+    process.exit(1);
 }
