+++ plugins/blog/src/app_admin/blog/categories/page.tsx
